+++ package.json
@@ -1,42 +1,45 @@
 {
-  "name": "stsloanertracker",
-  "version": "1.0.0",
-  "description": "App to track loaner laptops for STS",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
-    "dev": "nodemon index.js",
-    "db": "docker-compose up -d mongo",
-    "dev-db": "npm run db -d && npm run dev"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cse330-fall-2021/creativeproject-485964-475411.git"
-  },
-  "author": "Jack Heuberger, Jack Zhai, Eric Fang",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/cse330-fall-2021/creativeproject-485964-475411/issues"
-  },
-  "homepage": "https://github.com/cse330-fall-2021/creativeproject-485964-475411#readme",
-  "dependencies": {
-    "@json2csv/plainjs": "^6.1.3",
-    "body-parser": "^1.19.0",
-    "bootstrap": "^5.1.3",
-    "bootstrap-icons": "^1.7.1",
-    "dayjs": "^1.11.9",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-handlebars": "^7.1.2",
-    "mongoose": "^6.0.13",
-    "mongoose-lean-virtuals": "^0.9.1"
-  },
-  "devDependencies": {
-    "browser-sync": "^2.29.1",
-    "nodemon": "^3.0.1",
-    "@types/body-parser": "^1.19.2",
-    "@types/express": "^4.17.17",
-    "tsc": "^2.0.4"
-  }
+    "name": "stsloanertracker",
+    "version": "1.0.0",
+    "description": "App to track loaner laptops for STS",
+    "main": "index.js",
+    "scripts": {
+        "build": "tsc && tscp",
+        "start": "npm run build && node dist/index.js",
+        "dev": "nodemon src/index.ts",
+        "db": "docker-compose up -d mongo",
+        "dev-db": "npm run db -d && npm run dev"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/cse330-fall-2021/creativeproject-485964-475411.git"
+    },
+    "author": "Jack Heuberger, Jack Zhai, Eric Fang",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/cse330-fall-2021/creativeproject-485964-475411/issues"
+    },
+    "homepage": "https://github.com/cse330-fall-2021/creativeproject-485964-475411#readme",
+    "dependencies": {
+        "@json2csv/plainjs": "^6.1.3",
+        "@typegoose/typegoose": "^11.4.1",
+        "body-parser": "^1.20.2",
+        "bootstrap": "^5.3.1",
+        "bootstrap-icons": "^1.10.5",
+        "dayjs": "^1.11.9",
+        "dotenv": "^16.3.1",
+        "express": "^4.18.2",
+        "express-handlebars": "^7.1.2",
+        "mongoose": "^7.4.2",
+        "mongoose-lean-virtuals": "^0.9.1"
+    },
+    "devDependencies": {
+        "@types/body-parser": "^1.19.2",
+        "@types/express": "^4.17.17",
+        "browser-sync": "^2.29.3",
+        "nodemon": "^3.0.1",
+        "tsc": "^2.0.4",
+        "typescript": "^5.1.6",
+        "typescript-cp": "^0.1.9"
+    }
 }
