+++ src/models/OS.ts
@@ -1,0 +1,4 @@
+export enum OS {
+    MACOS = 'MACOS',
+    WINDOWS = 'WINDOWS'
+}
