--- src/routes/index.js
@@ -1,17 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-const Record = require('../models/Record')
-
-router.get('/', async (req, res) => {
-  try {
-    const records = await Record.find({ isOpen: 'true' }).sort({ nextContactDate: 'ascending' }).lean()
-    res.render('index', {
-      records
-    })
-  } catch (err) {
-    console.log(err)
-  }
-})
-
-module.exports = router
